+++ Nos/Service/Analytics.swift
@@ -129,16 +129,16 @@ class Analytics {
         track("Database Statistics", properties: properties)
     }
     
-    func databaseCleanupStarted(inBackground: Bool) {
-        track("Database Cleanup Started", properties: ["in_background": inBackground])
+    func databaseCleanupStarted() {
+        track("Database Cleanup Started")
     }
     
     func databaseCleanupTaskExpired() {
         track("Database Cleanup Task Expired")
     }
     
-    func databaseCleanupCompleted() {
-        track("Database Cleanup Completed")
+    func databaseCleanupCompleted(duration: TimeInterval) {
+        track("Database Cleanup Completed", properties: ["duration": duration])
     }
     
     func logout() {